requirements.txt
fastapi
uvicorn
requests
pypdf
python-multipart
pytesseract
pdf2image
pillow